requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==6.0.0
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
asyncio-atexit==1.0.1
attrs==24.2.0
autogen-agentchat==0.4.0.dev2
autogen-core==0.4.0.dev2
blinker==1.9.0
cachetools==6.2.4
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.3.1
comm==0.2.2
debugpy==1.8.7
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
executing==2.1.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.46
grpcio==1.62.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.4.0
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.6
jiter==0.6.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
multidict==6.1.0
narwhals==2.15.0
nest-asyncio==1.6.0
numpy==2.4.1
openai==0.28.1
opentelemetry-api==1.27.0
packaging==24.1
pandas==2.3.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==3.20.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==0.19.1
pytz==2025.2
pyzmq==26.2.0
referencing==0.37.0
regex==2024.9.11
requests==2.32.3
rpds-py==0.30.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
SpeechRecognition==3.14.5
stack-data==0.6.3
streamlit==1.53.1
streamlit_mic_recorder==0.0.8
tenacity==9.1.2
tiktoken==0.8.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.3
urllib3==2.2.3
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.16.0
zipp==3.20.2
